Fully working Basic Game
This is the first basic game fully working.
The avatar jumps triggered by a key pressed, between 1 and 8; each one gives a different jump intensity to reach te proper level.

The player falls down when a wrong key is pressed or when you press nothing.

When the player falls away the game finish and you can restart it.