Code modulation, pitch det. bugfix
The code is being better organized.

The pitch detector object is now declared as global variable in order to create only 1 object even if the game is restarted.

In the index.html file the index.js file is declared at the end because it needs all the other modules variables to be already defined.